+++ backend/src/main/java/edonymyeon/backend/post/application/PostService.java
@@ -1,11 +1,6 @@
 package edonymyeon.backend.post.application;
 
-import static edonymyeon.backend.global.exception.ExceptionInformation.MEMBER_ID_NOT_FOUND;
-import static edonymyeon.backend.global.exception.ExceptionInformation.POST_ID_NOT_FOUND;
-import static edonymyeon.backend.global.exception.ExceptionInformation.POST_MEMBER_FORBIDDEN;
-
 import edonymyeon.backend.global.exception.EdonymyeonException;
-import edonymyeon.backend.global.exception.ExceptionInformation;
 import edonymyeon.backend.image.ImageFileUploader;
 import edonymyeon.backend.image.domain.Domain;
 import edonymyeon.backend.image.postimage.domain.PostImageInfo;
@@ -14,33 +9,22 @@ import edonymyeon.backend.image.postimage.repository.PostImageInfoRepository;
 import edonymyeon.backend.member.application.dto.MemberIdDto;
 import edonymyeon.backend.member.domain.Member;
 import edonymyeon.backend.member.repository.MemberRepository;
-import edonymyeon.backend.post.application.dto.GeneralFindingCondition;
-import edonymyeon.backend.post.application.dto.GeneralPostInfoResponse;
-import edonymyeon.backend.post.application.dto.GeneralPostsResponse;
 import edonymyeon.backend.post.application.dto.PostModificationRequest;
 import edonymyeon.backend.post.application.dto.PostRequest;
 import edonymyeon.backend.post.application.dto.PostResponse;
-import edonymyeon.backend.post.application.dto.ReactionCountResponse;
-import edonymyeon.backend.post.application.dto.SpecificPostInfoResponse;
-import edonymyeon.backend.post.application.dto.WriterDetailResponse;
 import edonymyeon.backend.post.domain.Post;
 import edonymyeon.backend.post.repository.PostRepository;
-import edonymyeon.backend.post.repository.PostSpecification;
 import edonymyeon.backend.thumbs.application.ThumbsService;
-import edonymyeon.backend.thumbs.dto.AllThumbsInPostResponse;
-import edonymyeon.backend.thumbs.dto.ThumbsStatusInPostResponse;
-import java.util.Collections;
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Slice;
-import org.springframework.data.domain.Sort.Direction;
-import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
+
+import java.util.Collections;
+import java.util.List;
+import java.util.Objects;
+
+import static edonymyeon.backend.global.exception.ExceptionInformation.*;
 
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
@@ -164,94 +148,5 @@ public class PostService {
                 imageFileUploader.uploadFiles(request.newImages()));
         post.updateImages(updatedPostImageInfos);
         postImageInfoRepository.saveAll(updatedPostImageInfos.getPostImageInfos());
-    }
-
-    public List<GeneralPostInfoResponse> findPostsByPagingCondition(final GeneralFindingCondition generalFindingCondition) {
-        PageRequest pageRequest = convertConditionToPageRequest(generalFindingCondition);
-        final Slice<Post> foundPosts = postRepository.findAllBy(pageRequest);
-        return foundPosts
-                .map(post -> GeneralPostInfoResponse.of(post, domain.getDomain()))
-                .toList();
-    }
-
-    private static PageRequest convertConditionToPageRequest(final GeneralFindingCondition generalFindingCondition) {
-        try {
-            return PageRequest.of(
-                    generalFindingCondition.getPage(),
-                    generalFindingCondition.getSize(),
-                    switch (generalFindingCondition.getSortDirection()) {
-                        case ASC -> Direction.ASC;
-                        case DESC -> Direction.DESC;
-                    },
-                    generalFindingCondition.getSortBy().getName()
-            );
-        } catch (IllegalArgumentException e) {
-            throw new EdonymyeonException(ExceptionInformation.POST_INVALID_PAGINATION_CONDITION);
-        }
-    }
-
-    public SpecificPostInfoResponse findSpecificPost(final Long postId, final MemberIdDto memberIdDto) {
-        final Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new EdonymyeonException(POST_ID_NOT_FOUND));
-
-        final Optional<Member> member = memberRepository.findById(memberIdDto.id());
-
-        final ReactionCountResponse reactionCountResponse = new ReactionCountResponse(
-                0, // TODO: 조회수 기능 구현 필요
-                0, // TODO: 댓글 수 기능 구현 필요
-                0 // TODO: 스크랩 기능 구현 필요
-        );
-        final AllThumbsInPostResponse allThumbsInPost = thumbsService.findAllThumbsInPost(postId);
-        final WriterDetailResponse writerDetailResponse = getWriterResponse(post.getMember());
-
-        if (member.isEmpty()) {
-            return SpecificPostInfoResponse.of(
-                    post,
-                    allThumbsInPost,
-                    writerDetailResponse,
-                    reactionCountResponse,
-                    domain.getDomain()
-            );
-        }
-
-        final ThumbsStatusInPostResponse thumbsStatusInPost = thumbsService.findThumbsStatusInPost(memberIdDto, postId);
-
-        return SpecificPostInfoResponse.of(
-                post,
-                allThumbsInPost,
-                writerDetailResponse,
-                reactionCountResponse,
-                thumbsStatusInPost,
-                member.get(),
-                domain.getDomain()
-        );
-    }
-
-    private WriterDetailResponse getWriterResponse(final Member member) {
-        if (Objects.isNull(member.getProfileImageInfo())) {
-            return new WriterDetailResponse(
-                    member.getId(),
-                    member.getNickname(),
-                    null
-            );
-        }
-        return new WriterDetailResponse(
-                member.getId(),
-                member.getNickname(),
-                domain.getDomain() + member.getProfileImageInfo().getStoreName()
-        );
-    }
-
-    public GeneralPostsResponse searchPosts(final String searchWord, final GeneralFindingCondition generalFindingCondition) {
-        final Specification<Post> searchResults = PostSpecification.searchBy(searchWord);
-        final PageRequest pageRequest = convertConditionToPageRequest(generalFindingCondition);
-
-        final Slice<Post> foundPosts = postRepository.findAll(searchResults, pageRequest);
-
-        List<GeneralPostInfoResponse> posts = foundPosts.stream()
-                .map(post -> GeneralPostInfoResponse.of(post, domain.getDomain()))
-                .toList();
-
-        return new GeneralPostsResponse(posts);
     }
 }
